test(paxos): isolate async_write failpoint tests into the integration binary
The async_write failpoint tests lived as #[cfg(test)] unit tests in
src/storage/mod.rs, so they ran in the lib test binary alongside non-failpoint
unit tests. cargo runs a binary's tests concurrently and the fail registry is
process-global, so while one of these tests had paxos_toolkit::storage::async_write
armed, a sibling test doing a batch_async write (e.g.
append_after_full_trim_survives_reopen's trim) could trip it and fail
intermittently under cargo test --all-features.

Move both tests (lost_compacted_index_after_trim_recovers_forward_only and the
metrics test failed_async_write_increments_labelled_failure_counter) into the
existing tests/failpoints.rs integration binary, rewritten against the public
API and the common helpers, and add a FAILPOINT_TEST_SERIAL parking_lot::Mutex
guard so the binary's failpoint tests are mutually serialized -- matching the
convention already used by tsoracle-openraft-toolkit, tsoracle-driver-file, and
tsoracle-tests. The lib test binary no longer arms any failpoint, so concurrent
unit tests can no longer be poisoned.
